Specify a digest for the nginx image
Due to the nature of Docker image tagging - that is tags are often
reused in backwards incompatible ways - it is important to find an
alternative method to pin down the version of the nginx image used to
server the application assets.

By using the immutable image shasum we can guarantee the same image is
used each time.