Add ember-cli-mirage
ember-cli-mirage is a client-side server to help build, test and demo
Ember applications. This frontend application will use the server in
development to rapidly implement features whilst also defining the
interface for the backend to satisfy. Additionally testing this
application becomes far easier, with the ability to use factories to
define the server's state under each test scenario.