Mock a server to return random todo items
Using ember-cli-mirage, this commit implements a basic server with
endpoints defined for the `todo` resource. Because the resource is
simple the default implementations for each is handed off to
ember-cli-mirage that is able to accommodate for the following:

 - List all `todo` records (with associations)
 - Get a specific `todo` record (with associations)
 - Create a new `todo` record
 - Update an existing `todo` record
 - Delete an existing `todo` record

The store is handled in memory, which is far more efficient under both
development and testing environments.

Additionally factories have been created for both the `todo` and
`comment` models assigning random (within some predefined parameters)
values to each of the models' fields giving some variety to the output.

When the server is started it will generate 5 `todo` records, each
having 3 associated `comment` records.