Add helm chart manifests
Requests to this service will be proxied through an nginx proxy server
sitting in front of Google Cloud Storage and configured using a
configmap. This is required to have request destined for the storage
backend handled correctly.

The nginx configuration, stored as a configmap, is deployed alongside
the nginx container and include the following directives:

 - Return 405 if request is not a HEAD or GET
 - Return 200 if request is destined for the /healthz route
 - Rewrite requests for / to /index.html
 - Set the headers for Google Cloud Storage
 - Hide headers coming from Google Cloud Storage

Additionally the manifests declare a single deployment (with 2 replicas)
running a single nginx container. The image for this container comes
straight from Dockerhub as there are no specific configurations
warranting a custom image.