Use a versioned API backend in production
When deployed this application will retrieve data from an API gateway
backend service using JSON:API as the serialisation layer. To allow for
the API to be versioned and for the infrastructure to route requests to
the correct backend, each request will be prefixed with a version
string and sent to a specifc API host.