Add pact contract testing tool
Contract testing is a way to ensure that services (such as an API
provider and a client) can communicate with each other. Without contract
testing, the only way to know that services can communicate is by using
expensive and brittle integration tests.

This ember addon exposes test helpers that may be used to create the
contract between this client and its provider (an API gateway). In
concrete terms, this addon allows for the creation of consumer-driven
assertions against a mock API generated with ember-cli-mirage, that are
then verified to guarantee the mock API behaves in the same way as the
real API.

The produced interactions file can then be used to verify the real API
independently.